Refactor boarding screen to separate SocialButton into its own file for better modularity and maintainability
