+++ lib/feature/auth/boarding_screen.dart
@@ -3,6 +3,7 @@ import 'package:google_fonts/google_fonts.dart';
 
 import '../../utils/routes/app_routes.dart';
 import '../../utils/const/images.dart';
+import 'social_button.dart';
 
 class BoardingScreen extends StatelessWidget {
   const BoardingScreen({super.key});
@@ -89,55 +90,6 @@ class BoardingScreen extends StatelessWidget {
             ],
           ),
         ],
-      ),
-    );
-  }
-}
-
-class SocialButton extends StatelessWidget {
-  final String icon;
-  final String text;
-  final Color color;
-  final Color textColor;
-  final VoidCallback onPressed;
-
-  const SocialButton({
-    super.key,
-    required this.icon,
-    required this.text,
-    required this.color,
-    required this.textColor,
-    required this.onPressed,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.symmetric(horizontal: 16.0, vertical: 8.0),
-      child: ElevatedButton(
-        style: ElevatedButton.styleFrom(
-          backgroundColor: color,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(30),
-          ),
-          padding: const EdgeInsets.symmetric(vertical: 14),
-        ),
-        onPressed: onPressed,
-        child: Row(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Image.asset(icon, height: 24),
-            const SizedBox(width: 10),
-            Text(
-              text,
-              style: GoogleFonts.poppins(
-                fontSize: 16,
-                fontWeight: FontWeight.w600,
-                color: textColor,
-              ),
-            ),
-          ],
-        ),
       ),
     );
   }
